+++ database/migrations/2019_05_29_151048_create_structures_owners_table.php
@@ -14,9 +14,9 @@ class CreateStructuresOwnersTable extends Migration
     public function up()
     {
         Schema::create('structures_owners', function (Blueprint $table) {
-            $table->increments('id');
-            $table->integer('user_id')->unsigned()->unique();
-            $table->integer('structure_id')->unsigned();
+            $table->bigIncrements('id');
+            $table->bigInteger('user_id')->unsigned()->unique();
+            $table->bigInteger('structure_id')->unsigned();
             $table->timestamps();
         });
