+++ database/migrations/2019_06_26_205447_create_company_data_table.php
@@ -14,7 +14,7 @@ class CreateCompanyDataTable extends Migration
     public function up()
     {
         Schema::create('company_data', function (Blueprint $table) {
-            $table->increments('id');
+            $table->bigIncrements('id');
 
             /* Booleans values */
             $table->boolean('accept_offers')->nullable()->default(null);
@@ -26,18 +26,18 @@ class CreateCompanyDataTable extends Migration
             $table->boolean('looking_for_accompaniment')->nullable()->default(null);
 
             /* Integer values */
-            $table->integer('share_capital')->nullable()->default(null);
-            $table->integer('employees_number')->nullable()->default(null);
-            $table->integer('clients_number')->nullable()->default(null);
-            $table->integer('turnover')->nullable()->default(null);
-            $table->integer('turnover_projection')->nullable()->default(null);
-            $table->integer('average_monthly_turnover')->nullable()->default(null);
-            $table->integer('logistic_cost')->nullable()->default(null);
-            $table->integer('marketing_cost')->nullable()->default(null);
-            $table->integer('banking_investment')->nullable()->default(null);
-            $table->integer('ebitda')->nullable()->default(null);
-            $table->integer('investment_sought')->nullable()->default(null);
-            $table->integer('gross_margin')->nullable()->default(null);
+            $table->bigInteger('share_capital')->nullable()->default(null);
+            $table->bigInteger('employees_number')->nullable()->default(null);
+            $table->bigInteger('clients_number')->nullable()->default(null);
+            $table->bigInteger('turnover')->nullable()->default(null);
+            $table->bigInteger('turnover_projection')->nullable()->default(null);
+            $table->bigInteger('average_monthly_turnover')->nullable()->default(null);
+            $table->bigInteger('logistic_cost')->nullable()->default(null);
+            $table->bigInteger('marketing_cost')->nullable()->default(null);
+            $table->bigInteger('banking_investment')->nullable()->default(null);
+            $table->bigInteger('ebitda')->nullable()->default(null);
+            $table->bigInteger('investment_sought')->nullable()->default(null);
+            $table->bigInteger('gross_margin')->nullable()->default(null);
 
 
             $table->timestamps();
